v3

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -21,35 +21,109 @@ namespace HotelFinder.API.Controllers
             _hotelService = hotelService;
         }
 
+        /// <summary>
+        /// Get All Hotels
+        /// </summary>
+        /// <returns></returns>
+
         [HttpGet]
-        public List<Hotel> Get()
+        public async Task<IActionResult> Get()
         {
-            return _hotelService.GetAllHotels();
+            var hotels= await _hotelService.GetAllHotels();
+            return Ok(hotels);// 200 olarak response döndür + body kısmına datayı ekle
         }
 
-        [HttpGet("{id}")]
-        public Hotel Get(int id)
+        /// <summary>
+        /// Get an hotel by id
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("[action]/{id}")]
+        public async Task<IActionResult> gethotelbyıd(int id)
         {
-            return _hotelService.GetHotelById(id);
+
+            var hotel = await _hotelService.GetHotelById(id);
+            if (hotel!=null)
+            {
+                
+                return Ok(hotel);//200 + data döndü
+            }
+
+            return NotFound();//404
         }
+
+        /// <summary>
+        /// Get an hotel by name
+        /// </summary>
+        /// <param name="name"></param>
+        /// <returns></returns>
+        [HttpGet]
+        [Route("[action]")]
+        public async Task<IActionResult> Gethotelbyname(string name)
+        {
+            var hotel = await _hotelService.GetHotelByName(name);
+            if (hotel!=null)
+            {
+                return Ok(hotel);
+            }
+
+            return NotFound();
+        }
+
+       
+
+        /// <summary>
+        /// create an hotel
+        /// </summary>
+        /// <param name="hotel"></param>
+        /// <returns></returns>
 
         [HttpPost]
-        public Hotel Post([FromBody]Hotel hotel)
+        public async Task<IActionResult> Post([FromBody]Hotel hotel)
         {
-            return _hotelService.CreateHotel(hotel);
+            
+                var createdHotel= await _hotelService.CreateHotel(hotel);
+                return CreatedAtAction("Get",new {id = createdHotel.Id},createdHotel);
+            
+
+            
         }
+
+        /// <summary>
+        /// update the hotel
+        /// </summary>
+        /// <param name="hotel"></param>
+        /// <returns></returns>
 
         [HttpPut]
-        public Hotel Put([FromBody] Hotel hotel)
+        public async Task<IActionResult> Put([FromBody] Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            if (await _hotelService.GetHotelById(hotel.Id) != null)
+            {
+                var upgradedHotel = await _hotelService.UpdateHotel(hotel);
+                return Ok(upgradedHotel);
+            }
+
+            return NotFound();
         }
 
+
+        /// <summary>
+        /// delete the hotel
+        /// </summary>
+        /// <param name="id"></param>
         [HttpDelete("{id}")]
 
-        public void Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            _hotelService.DeleteHotel(id);
+            if (await _hotelService.GetHotelById(id)!=null)
+            {
+                await _hotelService.DeleteHotel(id);
+                return Ok();
+            }
+
+            return NotFound();
         }
